feat: simplify attribute handling
Previously, we made sure to preserve duplicate attributes on elements,
and all operations on attributes needed to be aware of the possibility
of duplicates and return all values correctly. This was originally due
to a requirement imposed while specifying the behavior of this library.

It was determined that this wasn't actually important anymore, so we're
stripping out all this extra complexity and making the API behave as if
there is only ever a single canonical value for an attribute. If a
duplicate attribute is encountered, it overrides the value previously
set for that attribute.